requirements.txt
requests==2.3.0
python-dateutil==2.2
pytz==2014.4